sync gradle and handle nan error in float

+++ feature/card/presentation/src/main/java/com/esum/feature/card/presentation/component/LineBarState.kt
@@ -1,4 +1,4 @@
 package com.esum.feature.card.presentation.component
 
 
-data class LineBarState(val percentage: Double = 0.0, val count: String = "")
+data class LineBarState(val percentage: Float = 0.0f, val count: String = "")
